converted from type() factory to proper ctypes.Structure subclasses
